requirements.txt
streamlit
PyPDF2
faiss-gpu
python-dotenv
langchain
htmlTemplates
faiss-cpu
